src/functions-python/requirements.txt
azure-functions==1.19.0
azure-identity==1.15.0
pyodbc==5.0.1
requests==2.31.0
